+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - New schema extension 'x-json-deserializer' to specify a custom Jackson deserializer class.
+- New config parameter 'dateClassName' to specify a custom class to represent string schemas with format "date" in generated code.
+- New config parameter 'dateTimeClassName' to specify a custom class to represent string schemas with format "date-time" in generated code.
 
 ### Changed
 - Bump dependency com.networknt:json-schema-validator to v3.0.0.